2002-08-03  Mike Kestner <mkestner@speakeasy.net>
	* generator/Method.cs : Added IsGetter, IsSetter, ReturnType. Made
	  GenerateImport, GenerateBody public.
	* generator/Parameters.cs : Added Parameter::MarshalType
	* generator/Property.cs : Added logic to use methods instead of text
	  properties wherever possible.

svn path=/trunk/gtk-sharp/; revision=6391

+++ sample/Scribble.cs
@@ -29,7 +29,7 @@ namespace GtkSamples {
 			darea.ConfigureEvent += new ConfigureEventHandler (ConfigureEvent);
 			darea.MotionNotifyEvent += new MotionNotifyEventHandler (MotionNotifyEvent);
 			darea.ButtonPressEvent += new ButtonPressEventHandler (ButtonPressEvent);
-			darea.Events = EventMask.ExposureMask |
+			darea.Events = (int)EventMask.ExposureMask |
 							   EventMask.LeaveNotifyMask |
 							   EventMask.ButtonPressMask |
 							   EventMask.PointerMotionMask |
